requirements.txt
charset-normalizer==3.4.0
python-docx==1.1.2
